feat(repositories implementation): implementation for ktor repositories.

+++ shared/src/commonMain/kotlin/com/teka/chamaa_finance/dtos/AccountTypeDto.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class AccountTypeDTO(
-    val accountTypeId: Long = 0,
+    val accountTypeId: Int = 0,
     val accountName: String,
     val created_at: String?,
     val updated_at: String?,
